# [BEAM-13015, #21250] Optimize encoding to a ByteString