--- a/sdks/java/harness/jmh/build.gradle
+++ b/sdks/java/harness/jmh/build.gradle
@@ -31,6 +31,7 @@ configurations {
 }
 
 dependencies {
+    implementation project(path: ":sdks:java:core", configuration: "shadow")
     implementation project(path: ":sdks:java:harness", configuration: "shadow")
     implementation project(":runners:java-fn-execution")
     runtimeOnly library.java.slf4j_jdk14