@lukecwik 2022-07-19
Run GoPortable PreCommit